Move methods to GatewayAPI+Persistence.swift.
